+++ src/actor.rs
@@ -3,7 +3,6 @@ use tokio::sync::mpsc::{UnboundedReceiver, UnboundedSender, unbounded_channel};
 use tracing::Span;
 
 pub mod app;
-pub mod broadcast;
 pub mod config;
 pub mod config_watcher;
 pub mod drag_swap;
